Split Fal webhook mapping helpers
